Refactor NoteNameEditor component and helpers
