+++ src/components/NoteNameEditor/HeaderComponent/HeaderStyles.js
@@ -1,0 +1,21 @@
+import styled from 'styled-components';
+
+export const StyledHeader = styled.div`
+
+  position: absolute;
+  top: 0;
+  z-index: 3;
+  height: 1.5em;
+  width: 100%;
+  background-color: #aac;
+  padding: 0.25em;
+  color: black;
+  text-transform: uppercase;
+`;
+
+export const StyledButtonContainer = styled.div`
+  position: absolute;
+  top: 0;
+  right: 0;
+  text-align: right;
+`;
