+++ src/components/NoteNameEditor/NoteNameEditorStyles.js
@@ -1,0 +1,19 @@
+import styled from 'styled-components';
+
+export const StyledNoteNameEditor = styled.div`
+  position: fixed;
+  top: 0;
+  height: 4em;
+  width: 100%;
+  padding-top: 1.5em;
+  z-index: 2;
+  background-color: black;
+  color: white;
+`;
+
+export const StyledNoteEditorTextArea = styled.textarea`
+  width: 100%;
+  height: 100%;
+  padding: 0.5em;
+  background-color: black;
+`;
