+++ src/helpers/visNetworkHelper.js
@@ -1,0 +1,9 @@
+export class VisNetworkHelper {
+  static getTargetNoteId(networkEvent) {
+    return networkEvent.nodes[0];
+  }
+
+  static clickedOnNote(networkEvent) {
+    return networkEvent.nodes.length > 0;
+  }
+}
