--- src/ui/spinner/trash/horizontal-loading-bar.html
@@ -1,5 +1,0 @@
-<!-- horizontal loading bar with moving ball animation -->
-<div id="movingBallG">
-	<div class="movingBallLineG"></div>
-	<div id="movingBallG_1" class="movingBallG"></div>
-</div>
